+++ packages/search-adapter/src/empathy/__tests__/request-mappers/empathy-queryable-request-mapper.spec.ts
@@ -1,45 +1,90 @@
+import { RelatedTag } from '@empathy/search-types';
 import { Container } from 'inversify';
-import { Dictionary } from '../../../types';
+import { Dictionary, QueryableRequest } from '../../../types';
+import { DEFAULT_EMPATHY_ADAPTER_CONFIG } from '../../config/empathy-adapter.config';
 import { DEPENDENCIES } from '../../container/container.const';
 import { RequestMapperContext } from '../../empathy-adapter.types';
+import { EmpathyRequestQueryMapper, EmpathyRequestRelatedTagsQueryMapper } from '../../mappers';
 import { EmpathyQueryableRequestMapper } from '../../mappers/request/empathy-queryable-request.mapper';
 
 const container = new Container();
-const mockedQueryMapper = { map: jest.fn((str: string) => str) };
-const emptyContext: RequestMapperContext = { feature: '', url: '', requestOptions: {} };
-container.bind(DEPENDENCIES.RequestMappers.Parameters.query)
-  .toConstantValue(mockedQueryMapper);
+
+const relatedTagsMapper = new EmpathyRequestRelatedTagsQueryMapper();
+const spiedRelatedTagsMapper = jest.spyOn(relatedTagsMapper, 'map');
+container.bind(DEPENDENCIES.RequestMappers.Parameters.query).toConstantValue(relatedTagsMapper);
+
+const queryMapper = new EmpathyRequestQueryMapper(DEFAULT_EMPATHY_ADAPTER_CONFIG);
+const spiedQueryMapper = jest.spyOn(queryMapper, 'map');
+container.bind(DEPENDENCIES.RequestMappers.Parameters.query).toConstantValue(queryMapper);
+
 let mapper: EmpathyQueryableRequestMapper;
+const emptyContext: RequestMapperContext = { feature: '', url: '', requestOptions: {} };
 
 beforeEach(() => {
   container.snapshot();
   mapper = container.resolve(EmpathyQueryableRequestMapper);
-  mockedQueryMapper.map.mockClear();
+  jest.clearAllMocks();
 });
 
-afterEach(() => {
-  container.restore();
-});
+afterEach(() => container.restore());
 
 it('Adds the query when passed', () => {
-  const rawRequest: Dictionary<string> = { a: '1', b: '2', query: 'shirt' };
+  const rawRequest: QueryableRequest & Dictionary<string> = { a: '1', b: '2', query: 'shirt' };
   const request: Dictionary<string> = {};
   const returnedRequest = mapper.map(rawRequest, request, emptyContext);
 
   expect(request).toEqual({ a: '1', b: '2', q: 'shirt' });
   expect(rawRequest).not.toBe(request);
   expect(request).toBe(returnedRequest);
-  expect(mockedQueryMapper.map).toHaveBeenCalledTimes(1);
+  expect(spiedQueryMapper).toHaveBeenCalledTimes(1);
 });
 
-it('Does not add the query when not passed', () => {
-  const rawRequest: Dictionary<string> = { a: '1', b: '2' };
+it('Query mapper is not called when query is empty', () => {
+  const rawRequest: QueryableRequest & Dictionary<string> = { a: '1', b: '2', query: '' };
   const request: Dictionary<string> = {};
   const returnedRequest = mapper.map(rawRequest, request, emptyContext);
 
-  expect(rawRequest).toEqual(request);
   expect(rawRequest).not.toBe(request);
   expect(request).toBe(returnedRequest);
-  expect(mockedQueryMapper.map).not.toHaveBeenCalled();
+  expect(spiedQueryMapper).not.toHaveBeenCalled();
+});
 
+it('Maps related tags to query', () => {
+  const relatedTags: RelatedTag[] = [
+    { tag: 'city', previous: 'lego', query: 'lego city', selected: true },
+    { tag: 'friends', previous: 'lego', query: 'lego friends', selected: true }
+  ];
+  // This const has an any type because TS don't understand the type QueryableRequest & Dictionary<string> :(
+  const rawRequest: any = {
+    query: 'lego',
+    relatedTags: relatedTags,
+    rows: '24'
+  };
+  const request: Dictionary<string> = {};
+  mapper.map(rawRequest, request, emptyContext);
+
+  expect(request.q).toBe('lego city friends');
+  expect(request.rows).toBeDefined();
+  expect(spiedQueryMapper).toHaveBeenCalledTimes(1);
+  expect(spiedRelatedTagsMapper).toHaveBeenCalledTimes(1);
+});
+
+it('Does not modify the query if no related tags are passed', () => {
+  const query = 'lego';
+  const rawRequest: QueryableRequest & Dictionary<string> = { query, origin: 'default' };
+  const request: Dictionary<string> = {};
+  mapper.map(rawRequest, request, emptyContext);
+
+  expect(request.q).toBe(query);
+  expect(request.origin).toBeDefined();
+  expect(spiedQueryMapper).toHaveBeenCalledTimes(1);
+  expect(spiedRelatedTagsMapper).toHaveBeenCalledTimes(1);
+});
+
+it('Does not call mappers if there is no query and neither related tags', () => {
+  const rawRequest: QueryableRequest & Dictionary<string> = { query: '' };
+  mapper.map(rawRequest, {}, emptyContext);
+
+  expect(spiedQueryMapper).toHaveBeenCalledTimes(0);
+  expect(spiedRelatedTagsMapper).toHaveBeenCalledTimes(0);
 });