+++ packages/search-adapter/src/empathy/__tests__/request-mappers/param-mappers/empathy-request-filters-mapper.spec.ts
@@ -1,13 +1,13 @@
 import { Facet, Filter, FilterModel, MultiSelect, SimpleFilter } from '@empathy/search-types';
 import { deepMerge } from '@empathybroker/deep-merge';
 import { Container } from 'inversify';
-import { DeepPartial, Dictionary } from '../../../types';
-import { EmpathyAdapterConfig } from '../../config/empathy-adapter-config.types';
-import { DEFAULT_EMPATHY_ADAPTER_CONFIG } from '../../config/empathy-adapter.config';
-import { DEPENDENCIES } from '../../container/container.const';
-import { RequestMapperContext } from '../../empathy-adapter.types';
-import { EmpathyRequestFiltersSolrSyntaxMapper } from '../../mappers/request/params/empathy-request-filters-solr-syntax.mapper';
-import { EmpathyRequestFiltersMapper } from '../../mappers/request/params/empathy-request-filters.mapper';
+import { DeepPartial, Dictionary } from '../../../../types';
+import { EmpathyAdapterConfig } from '../../../config/empathy-adapter-config.types';
+import { DEFAULT_EMPATHY_ADAPTER_CONFIG } from '../../../config/empathy-adapter.config';
+import { DEPENDENCIES } from '../../../container/container.const';
+import { RequestMapperContext } from '../../../empathy-adapter.types';
+import { EmpathyRequestFiltersSolrSyntaxMapper } from '../../../mappers/request/params/empathy-request-filters-solr-syntax.mapper';
+import { EmpathyRequestFiltersMapper } from '../../../mappers/request/params/empathy-request-filters.mapper';
 
 const container = new Container();
 const emptyContext: RequestMapperContext = { url: 'url', requestOptions: {}, feature: '' };