+++ packages/search-adapter/src/empathy/mappers/request/empathy-queryable-request.mapper.ts
@@ -1,19 +1,22 @@
-import { inject, injectable } from 'inversify';
-import { Dictionary } from '../../../types';
+import { injectable, multiInject } from 'inversify';
+import { Dictionary, QueryableRequest } from '../../../types';
 import { DEPENDENCIES } from '../../container/container.const';
-import { RequestMapper, RequestMapperContext } from '../../empathy-adapter.types';
+import { MapRequest, RequestMapper, RequestMapperContext } from '../../empathy-adapter.types';
+import { pipeMappers } from '../pipe-mappers';
 
 @injectable()
-export class EmpathyQueryableRequestMapper implements RequestMapper<Dictionary<string>, Dictionary<string>> {
+export class EmpathyQueryableRequestMapper implements RequestMapper<QueryableRequest & Dictionary<string>, Dictionary<string>> {
+  private readonly mapQuery: MapRequest<QueryableRequest, string>;
 
-  constructor(
-    @inject(DEPENDENCIES.RequestMappers.Parameters.query) private readonly queryMapper: RequestMapper<string, string>
-  ) {}
+  constructor(@multiInject(DEPENDENCIES.RequestMappers.Parameters.query) queryMapper: RequestMapper<QueryableRequest, string>[]) {
+    this.mapQuery = pipeMappers(...queryMapper);
+  }
 
-  map({ query, ...rest }: Dictionary<string>, request: Dictionary<string>, context: RequestMapperContext): Dictionary<string> {
+  map({ query, relatedTags = [], ...rest }: QueryableRequest & Dictionary<string>, request: Dictionary<string>,
+    context: RequestMapperContext): Dictionary<string> {
     return Object.assign(request, {
       ...rest,
-      q: query && this.queryMapper.map(query, query, context)
+      q: query && this.mapQuery({ query, relatedTags }, query, context)
     });
   }
 }