+++ packages/search-adapter/src/empathy/mappers/request/empathy-search-request.mapper.ts
@@ -1,6 +1,6 @@
-import { Filter, RelatedTag } from '@empathy/search-types';
+import { Filter } from '@empathy/search-types';
 import { injectable, multiInject } from 'inversify';
-import { Dictionary, SearchRequest } from '../../../types';
+import { Dictionary, QueryableRequest, SearchRequest } from '../../../types';
 import { DEPENDENCIES } from '../../container/container.const';
 import { MapRequest, RequestMapper, RequestMapperContext } from '../../empathy-adapter.types';
 import { EmpathySearchRequest } from '../../models';
@@ -8,11 +8,11 @@ import { pipeMappers } from '../pipe-mappers';
 
 @injectable()
 export class EmpathySearchRequestMapper implements RequestMapper<SearchRequest, EmpathySearchRequest> {
+  private readonly mapQuery: MapRequest<QueryableRequest, string>;
   private readonly mapFilters: MapRequest<Dictionary<Filter[]>, string[]>;
-  private readonly mapQuery: MapRequest<string, string>;
 
   constructor(
-    @multiInject(DEPENDENCIES.RequestMappers.Parameters.query) queryMapper: RequestMapper<string, string>[],
+    @multiInject(DEPENDENCIES.RequestMappers.Parameters.query) queryMapper: RequestMapper<QueryableRequest, string>[],
     @multiInject(DEPENDENCIES.RequestMappers.Parameters.filters) filtersMapper: RequestMapper<Dictionary<Filter[]>, string[]>[]
   ) {
     this.mapQuery = pipeMappers(...queryMapper);
@@ -23,14 +23,9 @@ export class EmpathySearchRequestMapper implements RequestMapper<SearchRequest, 
     context: RequestMapperContext): EmpathySearchRequest {
     return Object.assign<EmpathySearchRequest, Partial<EmpathySearchRequest>>(request, {
         ...rest,
-        q: this.mapQuery(query, this.applyRelatedTags(query, relatedTags), context),
+        q: query && this.mapQuery({ query, relatedTags }, query, context),
         filter: this.mapFilters(filters, [], context)
       }
     );
-  }
-
-  private applyRelatedTags(query: string, relatedTags: RelatedTag[]): string {
-    const relatedTagsValue = relatedTags.map(tag => tag.tag).join(' ');
-    return `${ query } ${ relatedTagsValue }`.trim();
   }
 }