+++ packages/search-adapter/src/empathy/mappers/request/params/empathy-request-query.mapper.ts
@@ -1,17 +1,18 @@
 import { inject, injectable } from 'inversify';
+import { QueryableRequest } from '../../../../types';
 import { EmpathyAdapterConfig, QueryConfig } from '../../../config/empathy-adapter-config.types';
 import { DEPENDENCIES } from '../../../container/container.const';
 import { RequestMapper } from '../../../empathy-adapter.types';
 
 @injectable()
-export class EmpathyRequestQueryMapper implements RequestMapper<string, string> {
+export class EmpathyRequestQueryMapper implements RequestMapper<QueryableRequest, string> {
   private readonly queryConfig: QueryConfig;
 
   constructor(@inject(DEPENDENCIES.config) config: EmpathyAdapterConfig) {
     this.queryConfig = config.mappings.query;
   }
 
-  map(_rawQuery: string, query: string): string {
+  map(_: QueryableRequest, query: string): string {
     return this.normalizeQuery(query);
   }