+++ packages/search-adapter/src/empathy/mappers/response/results/empathy-result.mapper.ts
@@ -52,7 +52,8 @@ export class EmpathyResultMapper implements ResponseMapper<EmpathyResult, Result
         value: sku,
         html: sku ? this.highlightWithQuery(sku, sku, context) : sku
       },
-      tagging: this.createResultTagging(rawResult, result.tagging || {}, context)
+      tagging: this.createResultTagging(rawResult, result.tagging || {}, context),
+      isWishlisted: rawResult.isWishlisted || false
     });
   }