+++ packages/search-adapter/src/empathy/models/entities/empathy-result.model.ts
@@ -13,4 +13,5 @@ export interface EmpathyResult {
   price?: number | string;
   rating?: string;
   url: string;
+  isWishlisted?: boolean;
 }